backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
google-generativeai
requests
beautifulsoup4
pydantic
psycopg2-binary
